+++ platform/android/src/style/layers/symbol_layer.hpp
@@ -35,6 +35,8 @@ public:
 
     jni::Object<jni::ObjectTag> getSymbolAvoidEdges(jni::JNIEnv&);
 
+    jni::Object<jni::ObjectTag> getSymbolZOrder(jni::JNIEnv&);
+
     jni::Object<jni::ObjectTag> getIconAllowOverlap(jni::JNIEnv&);
 
     jni::Object<jni::ObjectTag> getIconIgnorePlacement(jni::JNIEnv&);